Convert graph initializers into OrtValue Phase I (#23979)
### Description
This PR converts TensorProto graph initializers to TensorProto/OrtValue
pairs.
Currently, we only split the output for some optimizers to the above
pairs.
Eventually, we should be able to convert all initializers to OrtValues
on load.
Small weights will continue to be an exception as they are sometimes
required by ONNX inference functions.
Some graph API leaks to EPs so we are not able to remove it at present,
and this constrains our ability to convert everything at once.

### Motivation and Context
Lay Gound for proper layers separation. Eventually eliminate weights
copies in the EPs.