[EP ABI] Add ability to drop constant initializers for fused nodes specified in GetCapability (#25137)
### Description
- Add ability to drop constant initializers for fused nodes specified in
GetCapability.
- Rework how an EP specifies nodes that should be fused into one node
within GetCapability.
- Instead of passing the set of nodes as arguments to
`GraphSupportInfo_AddNodesToFuse()`, the EP creates an
`OrtNodeFusionOptions` object to specify the nodes and other relevant
options. This makes it easier to extend the API in the future since we
can't add more parameters to an existing function, but we can add more
functions that modify an options object.



### Motivation and Context
Add more functionality missing from GetCapability() in the EP ABI.