[EP ABI] Update OrtGraph to use new OrtValues stored in internal Graph (#25159)
### Description
Updates the `OrtGraph` implementation to take advantage of the work done
in PR https://github.com/microsoft/onnxruntime/pull/23979, which sets
the infrastructure to store initializers as `OrtValue` instances in the
`onnxruntime::Graph`.

There still needs to be second part to the [aforementioned
PR](https://github.com/microsoft/onnxruntime/pull/23979) to ensure that
all initializers are stored as `OrtValue`s in the Graph.



### Motivation and Context
<!-- - Why is this change required? What problem does it solve?
- If it fixes an open issue, please link to the issue here. -->