Allocator and data transfer support for plugin EP API (#25070)
### Description
<!-- Describe your changes. -->
Add allocator and data transfer infrastructure for plugin EP API

Allocators are created via the OrtEpFactory using OrtMemoryInfo that as
added to the OrtEpDevice instances the factory returns. This allows
allocators to be created outside of an inference session and shared.

When a library is loaded a default instance of each allocator is added
to the shared allocators if there is no existing allocator (e.g. user
provided custom allocator).
CreateSharedAllocator can be used to replace this default instance with
a user configured one. e.g. add an arena or provide other configuration
options that are passed through to the OrtEpFactory's CreateAllocator
function.
 
Similarly IDataTransfer is supported by the factory implementing
OrtDataTransferImpl, which will also enable data transfer outside of a
session. That will be added in a future PR as the synchronization
requirements need to be figured out and will affect the public API.

### Motivation and Context
<!-- - Why is this change required? What problem does it solve?
- If it fixes an open issue, please link to the issue here. -->

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>