add back linker flags "-z noexecstack" that were lost by accident (#25200)
### Description
<!-- Describe your changes. -->
Add back the linker option to make stack non-executable, which was
accidentally lost here:
https://github.com/microsoft/onnxruntime/pull/22646

This just adds back the option in the same place where it was.

### Motivation and Context
After upgrading to 1.22.0 we saw this warning:

OpenJDK 64-Bit Server VM warning: You have loaded library
/opt/vespa-deps/lib64/libonnxruntime.so.1.22.0 which might have disabled
stack guard. The VM will try to fix the stack guard now.
It's highly recommended that you fix the library with 'execstack -c
<libfile>', or link it with '-z noexecstack'.