Exclude EPContext Op from Common Subexpression Elimination graph optimization (#25296)
In the context of a model containing EPContext nodes, it's highly
unlikely that two EPContext nodes will produce the same results.
Furthermore, the EquivalenceClass constructor includes the node and all
its attributes in the hash calculation, which can be particularly
time-consuming when the "ep_cache_context" attribute contains a large
binary blob.

Therefore, we exclude EPContext op from CSE.