[webgpu] add usage of ReadonlyAllocator for WebGPU EP (#25690)
### Description

add usage of `ReadonlyAllocator` for WebGPU EP


### Motivation and Context

`ReadonlyAllocator` is added in #25348 to allow an EP to register a
separated allocator only use for initializers.

WebGPU EP already has different handling to initializers and
non-initializers, and this change makes WebGPU EP to use the preferred
way to deal with it. Now the allocator depends on the `OrtAllocatorType`
instead of the session initialization status.