[ABI] Migrate QNN EP for ABI Compatibility.
Re-implement QNN EP with ORT C ABI interfaces, including inheriting
OrtEpFactory and OrtEp, revising return status and logging mechanism.
Most importantly, all classes not provided in C ABI headers are not
allowed to be used after this migration, and therefore several parts are
either rewrite to achieve the same logic as before but not using any
prohibited classes or directly copy the implementations due to some are
too complex for rewrite.
Test: UTs are copied for ABI version but with ABI prefix in the test
suite name.