Move KleidiAI support detection to CMake logic. (#26178)
### Description
<!-- Describe your changes. -->

Use of the KleidiAI library is not supported on all platforms. Previously, this support detection was done in both
`tools/ci_build/build.py` and `cmake/CMakeLists.txt`. This change consolidates and simplifies the checks. They are now done in one place in the CMake logic.

Move setting of onnxruntime_target_platform into a separate file.

### Motivation and Context
<!-- - Why is this change required? What problem does it solve?
- If it fixes an open issue, please link to the issue here. -->

Clean up and simplify.