Quant tool: Don't store original outputs in MinMax Calibrator (#26321)
### Description
The MixMax calibrator stores all the outputs of each session run as
intermediate outputs. However, it only needs the augmented min and max
values. This causes the memory usage to unnecessarily grow with each
iteration of the calibration data, especially for llms which have large
outputs (logits and kv cache).

This PR fixes this by not storing the values from the model's original
outputs that are not needed for calibration.

### Motivation and Context
Improve memory usage.

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>